Move test bodies to the 4a and 4b files
